# Convert `as_df_row()` inputs to lists with elements of size 1