--- a/R/slice.R
+++ b/R/slice.R
@@ -194,6 +194,6 @@ vec_init <- function(x, n = 1L) {
 }
 
 # Used internally by `vec_rbind()`, but exported for testing
-vec_split_list <- function(x) {
-  .Call(vctrs_split_list, x)
+vec_split_along <- function(x) {
+  .Call(vctrs_split_along, x)
 }