--- a/src/slice.c
+++ b/src/slice.c
@@ -523,6 +523,186 @@ SEXP vctrs_as_index(SEXP i, SEXP n, SEXP names) {
   return vec_as_index(i, r_int_get(n, 0), names);
 }
 
+// Used in `as_df_row()` to turn `x` into a list where each element is size 1
+// containing the i-th element of `x`
+// [[ include("vctrs.h"); register() ]]
+SEXP vec_split_list(SEXP x) {
+  int nprot = 0;
+
+  R_len_t size = vec_size(x);
+
+  struct vctrs_proxy_info info = vec_proxy_info(x);
+  PROTECT_PROXY_INFO(&info, &nprot);
+
+  SEXP data = info.proxy;
+
+  SEXP index = PROTECT_N(r_int(0), &nprot);
+  int* p_index = INTEGER(index);
+
+  SEXP out = PROTECT_N(Rf_allocVector(VECSXP, size), &nprot);
+
+  // Holds a single result from slicing `x`
+  PROTECT_INDEX elt_prot_idx;
+  SEXP elt = R_NilValue;
+  PROTECT_WITH_INDEX(elt, &elt_prot_idx);
+  ++nprot;
+
+  // Fallback to `[` if the class doesn't implement a proxy. This is
+  // to be maximally compatible with existing classes.
+  if (OBJECT(x) && info.proxy_method == R_NilValue) {
+    if (info.type == vctrs_type_scalar) {
+      Rf_errorcall(R_NilValue, "Can't slice a scalar");
+    }
+
+    if (has_dim(x)) {
+      for (R_len_t i = 0; i < size; ++i) {
+        ++(*p_index);
+
+        elt = vec_slice_fallback(x, index);
+        REPROTECT(elt, elt_prot_idx);
+
+        if (ATTRIB(elt) == R_NilValue) {
+          elt = vec_restore(elt, x, R_NilValue);
+        }
+
+        SET_VECTOR_ELT(out, i, elt);
+      }
+
+      UNPROTECT(nprot);
+      return out;
+    }
+
+    PROTECT_INDEX call_prot_idx;
+    SEXP call = R_NilValue;
+    PROTECT_WITH_INDEX(call, &call_prot_idx);
+    ++nprot;
+
+    for (R_len_t i = 0; i < size; ++i) {
+      ++(*p_index);
+
+      call = Rf_lang3(fns_bracket, x, index);
+      REPROTECT(call, call_prot_idx);
+
+      elt = Rf_eval(call, R_GlobalEnv);
+      REPROTECT(elt, elt_prot_idx);
+
+      if (ATTRIB(elt) == R_NilValue) {
+        elt = vec_restore(elt, x, R_NilValue);
+      }
+
+      SET_VECTOR_ELT(out, i, elt);
+    }
+
+    UNPROTECT(nprot);
+    return out;
+  }
+
+  switch (info.type) {
+  case vctrs_type_null: {
+    UNPROTECT(nprot);
+    return R_NilValue;
+  }
+  case vctrs_type_logical:
+  case vctrs_type_integer:
+  case vctrs_type_double:
+  case vctrs_type_complex:
+  case vctrs_type_character:
+  case vctrs_type_raw:
+  case vctrs_type_list: {
+    if (has_dim(x)) {
+      SEXP names = PROTECT_N(Rf_getAttrib(x, R_DimNamesSymbol), &nprot);
+
+      SEXP row_names;
+      if (names != R_NilValue) {
+        row_names = VECTOR_ELT(names, 0);
+      }
+
+      PROTECT_INDEX new_names_prot_idx;
+      SEXP new_names = R_NilValue;
+      PROTECT_WITH_INDEX(new_names, &new_names_prot_idx);
+      ++nprot;
+
+      PROTECT_INDEX new_row_names_prot_idx;
+      SEXP new_row_names = R_NilValue;
+      PROTECT_WITH_INDEX(new_row_names, &new_row_names_prot_idx);
+      ++nprot;
+
+      for (R_len_t i = 0; i < size; ++i) {
+        ++(*p_index);
+
+        elt = vec_slice_shaped(info.type, data, index);
+        REPROTECT(elt, elt_prot_idx);
+
+        if (names != R_NilValue) {
+          new_names = Rf_shallow_duplicate(names);
+          REPROTECT(new_names, new_names_prot_idx);
+
+          new_row_names = slice_names(row_names, index);
+          REPROTECT(new_row_names, new_row_names_prot_idx);
+
+          SET_VECTOR_ELT(new_names, 0, new_row_names);
+
+          Rf_setAttrib(elt, R_DimNamesSymbol, new_names);
+        }
+
+        elt = vec_restore(elt, x, R_NilValue);
+
+        SET_VECTOR_ELT(out, i, elt);
+      }
+
+      UNPROTECT(nprot);
+      return out;
+    }
+
+    SEXP names = PROTECT_N(Rf_getAttrib(x, R_NamesSymbol), &nprot);
+
+    PROTECT_INDEX name_prot_idx;
+    SEXP name = R_NilValue;
+    PROTECT_WITH_INDEX(name, &name_prot_idx);
+    ++nprot;
+
+    for (R_len_t i = 0; i < size; ++i) {
+      ++(*p_index);
+
+      elt = vec_slice_base(info.type, data, index);
+      REPROTECT(elt, elt_prot_idx);
+
+      if (names != R_NilValue) {
+        name = slice_names(names, index);
+        REPROTECT(name, name_prot_idx);
+        r_poke_names(elt, name);
+      }
+
+      elt = vec_restore(elt, x, R_NilValue);
+
+      SET_VECTOR_ELT(out, i, elt);
+    }
+
+    UNPROTECT(nprot);
+    return out;
+  }
+
+  case vctrs_type_dataframe: {
+    for (R_len_t i = 0; i < size; ++i) {
+      ++(*p_index);
+
+      elt = df_slice(data, index);
+      REPROTECT(elt, elt_prot_idx);
+
+      elt = vec_restore(elt, x, R_NilValue);
+
+      SET_VECTOR_ELT(out, i, elt);
+    }
+
+    UNPROTECT(nprot);
+    return out;
+  }
+  default:
+    Rf_error("Internal error: Unexpected type `%s` for vector proxy in `vec_split_list()`",
+             vec_type_as_str(info.type));
+  }
+}
+
 void vctrs_init_slice(SEXP ns) {
   syms_vec_slice_fallback = Rf_install("vec_slice_fallback");
   fns_vec_slice_fallback = Rf_findVar(syms_vec_slice_fallback, ns);