… tests/testthat/test-slice.R