Closes #480 

This PR technically fixes #480, but I am hoping it can be improved, as I think it will be a bit slow in the long run.

It replaces the call to `r_as_list()` with an alternative, `as_list_row()`, that does something slightly different. It turns its input into a list such that each element of the result is of size 1. Meaning that `list(1:2, 1)` becomes `list(list(1:2), list(1))`. Another way to think of it is that it iterates through `x`, calling `vec_slice(x, i)` for each `i`, and places the result of the slice into the `i`th element of the output list. I think this is the most consistent way for `as_df_row()` to work.

I think `as_list_row()` is going to be somewhat slow though, as the repeated calls to `vec_slice()` might be a problem. Note that this isn't called for data frames (`as_df_row_impl()` early exits), but is called when the user does something like `vec_rbind(1:5)`, which might hurt performance. I'm open to other ideas!