@DavisVaughan 2019-07-24
We could probably remove `r_as_list()` now because I think this is the only place it was used