@DavisVaughan 2019-07-25
I've optimized it as much as I know how. Here are a few comparisons. We sacrifice a bit of speed when passing in vectors, but gain correct behavior with lists. Again, when in the most common case of working with data frames as input, we won't have any performance changes.

``` r
library(vctrs)
library(tibble)
library(rlang, warn.conflicts = FALSE)

# Will make a 2 row data frame
x <- set_names(1:5000)
y <- set_names(5001:10000, nm = names(x))
```

```r
# Can tell a performance difference between the two

# before
bench::mark(vec_rbind(x, y), iterations = 500)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(x, y)   4.13ms   4.91ms      195.     946KB     7.28

# after
bench::mark(vec_rbind(x, y), iterations = 500)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(x, y)   5.13ms   5.89ms      160.     985KB     8.40
```

```r
as_df_row <- vctrs:::as_df_row

# Performance loss is here :/

# before
bench::mark(as_df_row(x), iterations = 500)
#> # A tibble: 1 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 as_df_row(x)    292µs    322µs     3076.    90.8KB     31.1

# after
bench::mark(as_df_row(x), iterations = 500)
#> # A tibble: 1 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 as_df_row(x)    771µs    993µs      995.    90.8KB     18.2
```

```r
# But this works correctly now:

# before 
as_df_row(list(x = 1:2, y = 2))
#> Warning in format.data.frame(if (omit) x[seq_len(n0), , drop = FALSE] else
#> x, : corrupt data frame: columns will be truncated or padded with NAs
#>   x y
#> 1 1 2

# after
as_df_row(list(x = 1:2, y = 2))
#>      x y
#> 1 1, 2 2
```

```r
# And so does this:

# before
as_tibble(as_df_row(list(cars = mtcars)))
#> Tibble columns must have consistent lengths, only values of length one are recycled:
#> * Length 1: Requested with `.rows` argument
#> * Length 32: Column `cars`

# after
as_tibble(as_df_row(list(cars = mtcars)))
#> # A tibble: 1 x 1
#>   cars               
#>   <list>             
#> 1 <df[,11] [32 × 11]>
```