@lionel- 2019-07-25
Maybe a generic `r_as_list()` using the slice internals would solve the performance problems and be generally useful?