@DavisVaughan 2019-07-25
I think this works differently from `r_as_list()` though. If you pass `list(1, 2)` into `r_as_list()` I would expect it to return the results unchanged. But for `vec_rbind()` it would need to return `list(list(1), list(2))`. That's what I'm calling a "list row".