@lionel- 2019-07-25
> I think this works differently from r_as_list() though.

Good point, then it could be called `vec_split_list()`? This seems to make sense to me as an operation name because if you have an S3 list `<mylist: 1, 2, 3>`, you get `<list: <mylist: 1>, <mylist: 2>, <mylist: 3>>`.

> Note that this does use vec_slice_impl() now to avoid repeated calls to vec_as_index()

There's still redudant work done when calling `vec_slice_impl()` repeatedly with the same vector.