@DavisVaughan 2019-07-25
`vec_split_list()` makes sense to me. What you describe with the S3 object is exactly what should happen 👍 

> There's still redundant work done when calling vec_slice_impl() repeatedly with the same vector.

Yea I see that too. For how it will be used in `as_df_row()`, we could also make the assumption that it is not a shaped vector, and is not a data frame. I think one of the main speedups would be that we would be able to only call `vec_proxy_info()` once. And we wouldn't care about slicing names (`as_df_row()` takes care of that). I'll give it a shot.