@DavisVaughan 2019-07-26
One more example that didn't work before because of an inconsistency with `as.list()` that works now:

``` r
library(vctrs)
library(rlang)
as_df_row <- vctrs:::as_df_row

# before
as_df_row(set_names(factor(c("a", "b"))))
#> New names:
#> * `` -> ...1
#> * `` -> ...2
#>   ...1 ...2
#> 1    a    b

# after
as_df_row(set_names(factor(c("a", "b"))))
#>   a b
#> 1 a b
```