@DavisVaughan 2019-07-26
- `vec_split_list()` is fully implemented, and works in a generic enough way that it could be useful elsewhere. It is exported to the R side, but not to the user, and is fully tested.
- `vec_split_list()` basically eliminates the performance issues from before.

``` r
library(vctrs)
library(tibble)
library(rlang, warn.conflicts = FALSE)

# Will make a 2 row data frame
x <- set_names(1:5000)
y <- set_names(5001:10000, nm = names(x))
```

No more performance issues:

```r
# before
bench::mark(vec_rbind(x, y), iterations = 2500)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(x, y)   3.88ms   4.51ms      206.     946KB     7.96

# after
bench::mark(vec_rbind(x, y), iterations = 2500)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_rbind(x, y)   4.11ms   4.57ms      202.     985KB     7.81
```

It's still slightly slower than using `as.list()` like before, but it scales much better.

```r
as_df_row <- vctrs:::as_df_row

# before
bench::mark(as_df_row(x), iterations = 5000)
#> # A tibble: 1 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 as_df_row(x)    237µs    300µs     3301.    90.8KB     29.3

# after
bench::mark(as_df_row(x), iterations = 5000)
#> # A tibble: 1 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 as_df_row(x)    373µs    501µs     1957.    90.8KB     17.0
```

We can see what is taking time in `as_df_row()` by calling `vec_split_list()` by itself. Note that we need to unname `x` to get accurate timings, which is what `as_df_row()` does anyways, because that ends up removing all attributes from `x` so the repeated calls to `vec_restore()` run faster.

So `vec_split_list()` takes about 300us, and `as_df_row()` takes 500us. More investigation showed that `vec_unique_names()` was the other 200us.

```r
vec_split_list <- vctrs:::vec_split_list

bench::mark(vec_split_list(x), iterations = 10000)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_list(x)   1.05ms   1.49ms      657.    39.1KB     36.0

xx <- unname(x)
bench::mark(vec_split_list(xx), iterations = 10000)
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_list(xx)    202µs    302µs     3335.    39.1KB     29.9
```

Creating a row data frame from a classed object like a factor takes a little longer because we have to fall back to `[`. This example does not work with the current version because of how `as.list()` with factors works (the names are dropped from the factor and name repair has to kick in). I was impressed that `vec_split_list()` ends up being as fast / faster than `as.list.factor()`

```
# falls back to `[`
fctr <- factor(rep(letters, 100))
names(fctr) <- paste0("nm-", as.character(seq_along(fctr)))

bench::mark(as_df_row(fctr), iterations = 100)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 as_df_row(fctr)   8.69ms   10.7ms      93.8    57.3KB     55.1

bench::mark(vec_split_list(fctr), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_list(fctr)   10.8ms   12.4ms      70.0    20.4KB     52.8

bench::mark(as.list(fctr))
#> # A tibble: 1 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 as.list(fctr)   17.6ms   17.6ms      56.9    36.9KB    1536.
```

In my bench mark with `vec_split_list(x)`, I noted that it matters to unname `x` so it has no attributes. This is because we have to call `vec_restore()` on each sliced result from `x`, and this can be costly if there are any attributes. I don't think there is a good solution to this though.

```
z <- 1:5000
z_with_stuff <- structure(z, stuff = "stuff")

bench::mark(vec_split_list(z)) # vec_restore() early exits, no attributes
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_list(z)    236µs    280µs     3343.    58.7KB     29.5

bench::mark(vec_split_list(z_with_stuff)) # vec_restore() has to do work
#> # A tibble: 1 x 6
#>   expression                       min  median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_list(z_with_stuff)   781µs   965µs      988.    39.1KB     19.2
```

Having any attributes on `z` and then calling `as_df_row()` actually did not work before either, because `as.list()` nukes those attributes.

```r
# before
z_row <- as_df_row(z_with_stuff, quiet = TRUE)
z_row[[1]]
#> [1] 1

# after
z_row <- as_df_row(z_with_stuff, quiet = TRUE)
z_row[[1]]
#> [1] 1
#> attr(,"stuff")
#> [1] "stuff"
```

<sup>Created on 2019-07-26 by the [reprex package](https://reprex.tidyverse.org) (v0.2.1)</sup>